[AS7-6943] Cannot reliably uninstall bundles
