+++ server/src/main/java/org/jboss/as/server/deployment/Phase.java
@@ -472,9 +472,8 @@ public enum Phase {
     public static final int INSTALL_APPLICATION_CLIENT                  = 0x2010;
     public static final int INSTALL_MESSAGING_XML_RESOURCES             = 0x2030;
     public static final int INSTALL_BUNDLE_ACTIVATE                     = 0x2040;
-    public static final int INSTALL_BUNDLE_UNINSTALL                    = 0x2050;
-    public static final int INSTALL_WAB_SERVLETCONTEXT_SERVICE          = 0x2060;
-    public static final int INSTALL_PERSISTENCE_SERVICES                = 0x2070;
+    public static final int INSTALL_WAB_SERVLETCONTEXT_SERVICE          = 0x2050;
+    public static final int INSTALL_PERSISTENCE_SERVICES                = 0x2060;
     public static final int INSTALL_DEPLOYMENT_COMPLETE_SERVICE         = 0x2100;
 
     // CLEANUP
